# spec/openapi.yaml
openapi: 3.0.3
info:
  title: Seorilabs Platform API
  version: 1.0.0-draft
  description: |
    Seorilabs 공통 플랫폼 API.

    ## 계약 규칙

    - **이 문서가 계약의 source of truth다.** Go 타입도 TS 타입도 여기서 생성된 산출물이다
    - **`/v1`은 영구히 깨지지 않는다.** 마켓 배포된 구버전 SDK가 2~3년 산다
    - 필드 **추가만** 허용한다. 제거와 의미 변경은 `/v2`
    - 양방향 forward-compat — 서버는 요청의 미지 필드를 무시하고, SDK는 응답의 미지 필드를 무시한다
    - 모든 응답은 envelope로 감싼다. 성공은 `{ok:true, result}`, 실패는 `{ok:false, error}`
    - 요청 본문 상한 64KiB. 제어문자 거부. 허용 필드 외 어떤 키도 400

    상태: **D0 초안.** 스키마 세부는 각 단계에서 확정한다.
servers:
  - url: https://platform-api.seorilabs.example/v1
    description: 확정 필요 — P0 배포 후 실제 URL로 교체
...
paths:
...
  /iap/verify:
    post:
      tags: [iap]
      summary: 구매 검증과 지급
      description: |
        마켓 3종을 한 엔드포인트로 받는다. `platform`이 어느 마켓인지 정한다.

        **Apple은 NON_CONSUMABLE이 아닌 상품을 422로 거부한다.**
        **production과 sandbox 사이의 자동 fallback은 없다.**

        응답의 `granted`와 `alreadyGranted`는 항상 배타적이다. 둘 다
        false인 조합은 존재하지 않는다 — 불변식 2다.
      operationId: verifyIapPurchase
      parameters:
        - $ref: '#/components/parameters/SeoriApp'
        - $ref: '#/components/parameters/AppCheck'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [platform, productId, token]
              additionalProperties: false
              properties:
                platform:
                  type: string
                  enum: [google_play, app_store, apps_in_toss]
                productId: { $ref: '#/components/schemas/ProductId' }
                token:
                  type: string
                  maxLength: 4096
                  description: >
                    마켓이 발급한 증빙. Play는 purchaseToken, App Store는
                    transactionId, AppsInToss는 orderId다. 소유의 근거는
                    이 토큰이다 — 마켓이 이 계정에 발급했다는 사실 자체가
                    마켓이 내린 소유 판정이다. ADR 0010 참고.
      responses:
        '200': { $ref: '#/components/responses/VerifyPurchaseOk' }
        '400': { $ref: '#/components/responses/BadRequest' }
        '401': { $ref: '#/components/responses/Unauthorized' }
        '409': { $ref: '#/components/responses/Conflict' }
        '413': { $ref: '#/components/responses/PayloadTooLarge' }
        '415': { $ref: '#/components/responses/UnsupportedMediaType' }
        '422': { $ref: '#/components/responses/Unprocessable' }
        '429': { $ref: '#/components/responses/RateLimited' }
        '502': { $ref: '#/components/responses/ProviderError' }
        '503': { $ref: '#/components/responses/ServiceUnavailable' }
  /iap/entitlements:
    get:
      tags: [iap]
      summary: entitlement 목록 조회
      description: |
        마켓 SDK가 없어도 이 엔드포인트만으로 환불 projection을 반영할 수 있다.
      operationId: listEntitlements
      parameters:
        - $ref: '#/components/parameters/SeoriApp'
      responses:
        '200':
          description: 조회 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/EnvelopeOk'
                  - type: object
                    required: [result]
                    properties:
                      result:
                        type: object
                        required: [status, entitlements]
                        properties:
                          status:
                            type: string
                            enum: [verified]
                          entitlements:
                            type: array
                            items: { $ref: '#/components/schemas/EntitlementId' }
        '401': { $ref: '#/components/responses/Unauthorized' }
  /iap/account-references:
    post:
      tags: [iap]
      summary: 마켓 계정 참조 발급
      description: |
        Google과 Apple 신규 구매 전에 클라이언트가 받아 마켓 SDK에 넘긴다.

        HMAC keyring으로 생성한다. 첫 항목이 현재 키이고 나머지는 회전 검증용이다.
      operationId: getAccountReferences
      parameters:
        - $ref: '#/components/parameters/SeoriApp'
      responses:
        '200':
          description: 발급 성공
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/EnvelopeOk'
                  - type: object
                    properties:
                      result:
                        type: object
                        required: [googlePlayObfuscatedAccountId, appStoreAppAccountToken]
                        properties:
                          googlePlayObfuscatedAccountId:
                            type: string
                          appStoreAppAccountToken:
                            type: string
                            format: uuid
        '401': { $ref: '#/components/responses/Unauthorized' }

  # 실제 라우트는 /v1/iap/webhooks/apple 이다. App Store Connect의 알림
  # URL에 이 값을 그대로 등록한다. 경로가 틀리면 404가 나고 환불·취소
  # 알림이 유실된다 — 마켓은 우리 원장을 대신 지켜주지 않는다.
...
  /admin/apps/{appId}/ads/config:
    get:
      tags: [admin, ads]
      summary: 앱 광고 레지스트리 설정 조회
      security: [{ AdminOidc: [] }]
      parameters: [{ name: appId, in: path, required: true, schema: { type: string } }]
      responses:
        '200':
          description: suffix만 포함한 읽기 전용 설정
          content: { application/json: { schema: { allOf: [{ $ref: '#/components/schemas/EnvelopeOk' }, { type: object, required: [result], properties: { result: { $ref: '#/components/schemas/AdminAdsConfig' } } }] } } }
        '403': { $ref: '#/components/responses/Forbidden' }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/users/{puid}/ads:
    get:
      tags: [admin, ads]
      summary: 사용자 광고 정책과 감사 이력
      security: [{ AdminOidc: [] }]
      parameters: [{ name: puid, in: path, required: true, schema: { type: string } }]
      responses:
        '200':
          description: PII 없는 광고 정책
          content: { application/json: { schema: { allOf: [{ $ref: '#/components/schemas/EnvelopeOk' }, { type: object, required: [result], properties: { result: { $ref: '#/components/schemas/AdminUserAds' } } }] } } }
        '404': { $ref: '#/components/responses/NotFound' }
  /admin/ads/reward-claims:
    get:
      tags: [admin, ads]
      summary: 보상 claim 진단 목록
      security: [{ AdminOidc: [] }]
      parameters:
        - { name: appId, in: query, schema: { type: string } }
        - { name: provider, in: query, schema: { type: string, enum: [admob, apps_in_toss] } }
        - { name: state, in: query, schema: { type: string, enum: [accepted, confirmed, delivered, expired] } }
        - { name: assurance, in: query, schema: { type: string, enum: [pending, server_verified, client_confirmed] } }
        - { name: placement, in: query, schema: { type: string } }
        - { name: reference, in: query, description: 정확한 supportCode 또는 claim ID, schema: { type: string } }
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 500, default: 100 } }
      responses:
        '200':
          description: 원문 transaction과 사용자 식별자를 제외한 목록
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/EnvelopeOk'
                  - type: object
                    required: [result]
                    properties:
                      result:
                        type: object
                        required: [claims]
                        additionalProperties: false
                        properties:
                          claims: { type: array, items: { $ref: '#/components/schemas/AdRewardClaim' } }
  /admin/ads/suppressions/grant:
    post:
      tags: [admin, ads]
      summary: 운영자 광고 차단 추가
      security: [{ AdminOidc: [] }]
      parameters: [{ $ref: '#/components/parameters/AdminActor' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AdminAdsSuppressionRequest' } } } }
      responses:
        '200': { description: 반영 또는 이미 active인 멱등 성공, content: { application/json: { schema: { allOf: [{ $ref: '#/components/schemas/EnvelopeOk' }, { type: object, required: [result], properties: { result: { $ref: '#/components/schemas/AdminAdsSuppressionResult' } } }] } } } }
        '409': { $ref: '#/components/responses/Conflict' }
  /admin/ads/suppressions/revoke:
    post:
      tags: [admin, ads]
      summary: 운영자 광고 차단 회수
      security: [{ AdminOidc: [] }]
      parameters: [{ $ref: '#/components/parameters/AdminActor' }]
      requestBody: { required: true, content: { application/json: { schema: { $ref: '#/components/schemas/AdminAdsSuppressionRequest' } } } }
      responses:
        '200': { description: 회수 또는 동일 requestId 멱등 성공, content: { application/json: { schema: { allOf: [{ $ref: '#/components/schemas/EnvelopeOk' }, { type: object, required: [result], properties: { result: { $ref: '#/components/schemas/AdminAdsSuppressionResult' } } }] } } } }
        '409': { $ref: '#/components/responses/Conflict' }
components:
  securitySchemes:
    PlatformToken:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: '`/auth/session`이 발급한 플랫폼 세션 토큰'
    PubSubOidc:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Pub/Sub push subscription이 붙이는 Google OIDC 토큰
    AdminOidc:
      type: http
      scheme: bearer
      bearerFormat: Google OIDC ID token
      description: |
        Cloud Run Admin audience를 대상으로 발급한 Google OIDC ID token.
        GET은 read 또는 write allowlist, mutation은 write allowlist만 허용한다.
  parameters:
    SeoriApp:
      name: X-Seori-App
      in: header
      required: true
      description: |
        앱 식별자. **권한이 아니라 어느 프로젝트로 검증할지 고르는 힌트다.**
        헤더를 바꿔도 토큰의 aud 불일치로 거부된다.
      schema:
        type: string
        pattern: '^[a-z0-9][a-z0-9-]{0,63}$'
...
    AppCheck:
      name: X-Firebase-AppCheck
      in: header
      description: |
        레지스트리 `require_app_check`가 true인 앱에서만 필수다.

        1단계는 optional로 시작한다. Godot에 App Check SDK가 없고
        replay 방지(consume)가 Go SDK에 없다.
      schema: { type: string }
...
    AdminActor:
      name: X-Seori-Actor
      in: header
      required: false
      description: |
        작업을 승인한 GitHub login. 권한 판단에는 쓰지 않으며 형식만 검증한다.
        없거나 형식이 다르면 OIDC principal의 SHA-256 참조를 저장한다.
        grant, revoke, 최초 sandbox reset POST와 close-not-started의 멱등 payload
        일부이므로 unknown 결과를 같은 원 endpoint로 재시도할 때는 최초 값을 그대로
        보내야 한다. reset resume는 저장된 최초 payload를 사용한다.
      schema:
        type: string
        pattern: '^[A-Za-z0-9-]{1,39}$'
  schemas:
    EnvelopeOk:
      type: object
      required: [ok]
      properties:
        ok:
          type: boolean
          enum: [true]
    EnvelopeError:
      type: object
      required: [ok, error]
      properties:
        ok:
          type: boolean
          enum: [false]
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: snake_case 에러 코드. 전체 목록은 docs/03-architecture/iap.md
              example: purchase_owned_by_another_user
            message:
              type: string
              description: 한국어 메시지. 클라이언트가 이 문자열로 분기하지 않는다
...
    ProductId:
      type: string
      maxLength: 128
      pattern: '^[A-Za-z0-9._-]{1,128}$'
    EntitlementId:
      type: string
      maxLength: 128
      pattern: '^[A-Za-z0-9._-]{1,128}$'
...
    AdsPolicy:
      type: object
      required: [appUsesAds, adsEnabled, disabledBy, checkedAt]
      additionalProperties: false
      properties:
        appUsesAds: { type: boolean }
        adsEnabled: { type: boolean }
        disabledBy: { type: array, items: { type: string, enum: [operator, ad_free] } }
        checkedAt: { type: string, format: date-time }
    AdReward:
      type: object
      required: [key, amount]
      additionalProperties: false
      properties:
        key: { type: string, pattern: '^[a-z0-9][a-z0-9_-]{0,127}$' }
        amount: { type: integer, minimum: 1 }
...
    AdRewardClaim:
      type: object
      required: [claimId, appId, placement, provider, clientPlatform, reward, state, assurance, createdAt, expiresAt]
      additionalProperties: false
      properties:
        claimId: { type: string }
        appId: { type: string }
        placement: { type: string }
        provider: { type: string, enum: [admob, apps_in_toss] }
        clientPlatform: { type: string, enum: [android, ios, apps_in_toss] }
        reward: { $ref: '#/components/schemas/AdReward' }
        state: { type: string, enum: [accepted, confirmed, delivered, expired] }
        assurance: { type: string, enum: [pending, server_verified, client_confirmed] }
        createdAt: { type: string, format: date-time }
        confirmedAt: { type: string, format: date-time }
        acknowledgedAt: { type: string, format: date-time }
        expiresAt: { type: string, format: date-time }
        admobSsv:
          type: object
          additionalProperties: false
          required: [customData, userId]
          properties:
            customData: { type: string }
            userId: { type: string }
...
    AdminAdsProviderConfig:
      type: object
      additionalProperties: false
      properties:
        androidAdUnitSuffix: { type: string }
        iosAdUnitSuffix: { type: string }
        adGroupSuffix: { type: string }
        rewardItem: { type: string }
        rewardAmount: { type: integer, minimum: 0 }
    AdminAdsRewardRange:
      type: object
      required: [key, min_amount, max_amount]
      additionalProperties: false
      properties:
        key: { type: string }
        min_amount: { type: integer, minimum: 1 }
        max_amount: { type: integer, minimum: 1 }
    AdminAdsConfig:
      type: object
      required: [appId, providers, placements, registrySyncedAt]
      additionalProperties: false
      properties:
        appId: { type: string }
        providers: { type: array, items: { type: string, enum: [admob, apps_in_toss] } }
        registrySyncedAt: { type: string, format: date-time }
        placements:
          type: array
          items:
            type: object
            required: [id, format, providers, dailyLimit, cooldownSeconds]
            additionalProperties: false
            properties:
              id: { type: string }
              format: { type: string, enum: [rewarded, interstitial] }
              providers:
                type: object
                additionalProperties: { $ref: '#/components/schemas/AdminAdsProviderConfig' }
              reward: { $ref: '#/components/schemas/AdminAdsRewardRange' }
              dailyLimit: { type: integer, minimum: 1 }
              cooldownSeconds: { type: integer, minimum: 0 }
    AdminAdsSuppressionRecord:
      type: object
      required: [requestId, appId, platformUserId, actorLogin, reason, operation, applied, createdAt]
      additionalProperties: false
      properties:
        requestId: { type: string }
        grantRequestId: { type: string }
        appId: { type: string }
        platformUserId: { type: string }
        actorLogin: { type: string }
        reason: { $ref: '#/components/schemas/AdminMutationReason' }
        operation: { type: string, enum: [grant, revoke] }
        applied: { type: boolean }
        createdAt: { type: string, format: date-time }
    AdminUserAds:
      type: object
      required: [appId, platformUserId, supportCode, isAnonymous, authType, lastSeenAt, policy, auditHistory]
      additionalProperties: false
      properties:
        appId: { type: string }
        platformUserId: { type: string }
        supportCode: { type: string }
        isAnonymous: { type: boolean }
        authType: { type: string, enum: [firebase, firebase_bridge, apps_in_toss, anonymous] }
        lastSeenAt: { type: string, format: date-time }
        policy: { $ref: '#/components/schemas/AdsPolicy' }
        auditHistory: { type: array, items: { $ref: '#/components/schemas/AdminAdsSuppressionRecord' } }
    AdminAdsSuppressionResult:
      type: object
      required: [applied, requestId]
      additionalProperties: false
      properties:
        applied: { type: boolean }
        requestId: { type: string }
        activeGrantRequestId: { type: string }
    AdminAdsSuppressionRequest:
      type: object
      required: [requestId, appId, platformUserId, reason, confirmation]
      additionalProperties: false
      properties:
        requestId: { type: string, maxLength: 128 }
        appId: { type: string }
        platformUserId: { type: string }
        grantRequestId: { type: string, maxLength: 128 }
        reason: { $ref: '#/components/schemas/AdminMutationReason' }
        confirmation: { type: string, maxLength: 300 }
...
    VerifyPurchaseOutcome:
      type: object
      required: [status, entitlementId, entitlements]
      properties:
        status:
          type: string
          enum: [verified, pending, revoked]
        entitlementId:
          $ref: '#/components/schemas/EntitlementId'
        granted:
          type: boolean
          description: |
            **granted와 alreadyGranted는 항상 배타적이다.**
            둘 다 false인 조합은 존재하지 않는다. status가 verified일 때만 존재한다.
        alreadyGranted:
          type: boolean
        entitlements:
          type: array
          items: { $ref: '#/components/schemas/EntitlementId' }
        completion:
          $ref: '#/components/schemas/Completion'
        blockedBySandboxReset:
          type: boolean
          description: sandbox 초기화로 재지급이 차단된 경우에만 존재한다
    Completion:
      type: object
      required: [action]
      properties:
        action:
          type: string
          enum:
            - none
            - retry_server_completion
            - apps_in_toss_complete_product_grant
            - app_store_sync_after_sandbox_reset
          description: |
            retry_server_completion은 **지급이 이미 커밋된 상태**를 뜻한다.
            마켓 완료 호출만 실패했고 서버 outbox가 재시도한다. 롤백하지 않는다.
        orderId:
          type: string
          description: apps_in_toss_complete_product_grant일 때만
...
    AdminMutationReason:
      type: string
      description: PII나 결제 자격정보를 저장하지 않는 고정 운영 사유 코드
      enum:
        - customer_support_compensation
        - incorrect_grant_correction
        - incident_recovery
        - internal_validation
...
  responses:
    BadRequest:
      description: 요청 형식 오류. 허용 필드 외 키가 있어도 여기로 온다
      content:
        application/json:
          schema: { $ref: '#/components/schemas/EnvelopeError' }
    Unauthorized:
      description: 인증 실패
      content:
        application/json:
          schema: { $ref: '#/components/schemas/EnvelopeError' }
    NotFound:
      description: 대상을 찾지 못함
      content:
        application/json:
          schema: { $ref: '#/components/schemas/EnvelopeError' }
    Forbidden:
      description: 권한 부족, 앱 paused, anonymous 또는 관리 allowlist 불일치
      content:
        application/json:
          schema: { $ref: '#/components/schemas/EnvelopeError' }
    Conflict:
      description: 다른 소유자의 구매이거나 replay 불일치
      content:
        application/json:
          schema: { $ref: '#/components/schemas/EnvelopeError' }
    Unprocessable:
      description: 구매가 유효하지 않거나 상품 타입이 맞지 않음
      content:
        application/json:
          schema: { $ref: '#/components/schemas/EnvelopeError' }
    PayloadTooLarge:
      description: 본문 64KiB 초과
      content:
        application/json:
          schema: { $ref: '#/components/schemas/EnvelopeError' }
    UnsupportedMediaType:
      description: Content-Type이 application/json이 아님
      content:
        application/json:
          schema: { $ref: '#/components/schemas/EnvelopeError' }
    RateLimited:
      description: 한도 초과
      content:
        application/json:
          schema: { $ref: '#/components/schemas/EnvelopeError' }
    ProviderError:
      description: 마켓 API 오류
      content:
        application/json:
          schema: { $ref: '#/components/schemas/EnvelopeError' }
    ServiceUnavailable:
      description: 설정 오류이거나 일시적 처리 불가
      content:
        application/json:
          schema: { $ref: '#/components/schemas/EnvelopeError' }
...
    VerifyPurchaseOk:
      description: 검증 완료
      content:
        application/json:
          schema:
            allOf:
              - $ref: '#/components/schemas/EnvelopeOk'
              - type: object
                required: [result]
                properties:
                  result:
                    $ref: '#/components/schemas/VerifyPurchaseOutcome'
...
security:
  - PlatformToken: []